requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
gunicorn==22.0.0
python-multipart==0.0.9
azure-cognitiveservices-speech==1.38.0
